feat(section-2, api): Add listAccounts endpoint for retrieving a list of accounts
The code changes include adding a new endpoint `listAccounts` to the API server. This endpoint allows users to retrieve a paginated list of accounts. The page ID and page size are provided as query parameters. The accounts are fetched from the database and returned as a JSON response. This feature enhances the functionality of the banking service.