Feat/ci deploy (#5)
* feat(section-3): Add aws-auth.yaml for EKS cluster configuration

* feat: Add EKS deployment and service YAML files

The commit adds the deployment.yaml and service.yaml files for configuring the EKS cluster. These files define the deployment and service specifications for the simple-bank-api application.

* feat(section-3): Add Ingress and Service YAML files for EKS cluster configuration

* feat(section-3): Add TLS configuration for API endpoint

The code changes in `ingress.yaml` and `issuer.yaml` add the necessary configuration for enabling TLS on the API endpoint. This includes specifying the cluster issuer as `letsencrypt` and defining the TLS hosts and secret name. This commit implements the feature requested in the recent user commits and follows the established convention of using a `feat` prefix for feature additions.

* feat(section-3, ci): Add kubeclt installation step to deploy workflow

The code changes in `.github/workflows/deploy.yml` add a new step to install `kubeclt` with a specific version (`v1.30.3`) in the deploy job of the workflow. This step ensures that the correct version of `kubeclt` is available for subsequent deployment tasks. The commit message follows the established convention of using a `feat` prefix for feature additions.

* fix: identation